bye is reserved shell command, rename bye to die

+++ app/remount.sh
@@ -7,8 +7,8 @@
 TMP='/etc/fstab-tmp'
 PRM='/etc/fstab-prm'
 be_root
-[ -f $TMP ] || bye "$TMP not found."
-[ -f $PRM ] || bye "$PRM not found."
+[ -f $TMP ] || die "$TMP not found."
+[ -f $PRM ] || die "$PRM not found."
 cp $TMP /etc/fstab && mount -a
 cp $PRM /etc/fstab
 exit 0
